Add support for host elements in directives
